HACK-820 Add Tradeshift ALM and check name
The new plugin is missing some of the configuration options that were used in version 1.3 and were passed in via the scanner. Without these options the sonar build job will not complete until manual configuration is done on a per repo basis.

This change hardcodes our configuartion into the plugin.