Add support for Sonarqube 8.5.0
Sonarqube 8.5.0 removed the `BranchSupport.Branch` class which breaks backwards-compatibility with previous Sonarqube versions, and requires updating how the ComponentKey is checked and generated during the creation of the analysis report. This version of Sonarqube also moves the bundled plugins into their own directory, so the build classpath has been updated to exclude those plugins from being able to be part of the compilation path, and updates the class-loading unit tests to reference a plugin class and file path that is still available in the newer Sonarqube version.

Includes additional Web Service parameters introduced in Sonarqube 8.4.0 for configuring Github ALM integration so that these values can be persisted when provided, even though the plugin does not actually require them.